:sparkles: NEW: Add root templates directory
Added new directory called templates/ to hold base files for each
cookiecutter template. They still require some work and this is an
initial commit but they are intended to be symlinked or copied over to
the cookiecutter template using the pre-generation hook so that these
files are central and only require modification in a single place
avoiding duplication.